fix: resolve integration test failures across management and redteam
- Topics.Delete/ForceDelete: handle JSON string response (not object)
- Topics.ForceDelete: fix path /topic/{id}/force → /topic/force/{id}
- Topics.Update integration test: include required topic_name field
- OAuth.GetToken: accept OAuthTokenRequest (caller-provided client_id)
- Profiles.GetByName/DlpProfiles.Get: list+filter (no GET-by-name endpoint)
- Profiles.Update/Topics.Update: fix path to include /uuid/ segment
- redteam: TargetMetadata.RequestTimeout *int → *float64 (spec: number)
- OAuthClient: add ClientID() accessor for token request body

+++ aisec/redteam/models.go
@@ -813,7 +813,7 @@ type TargetMetadata struct {
 	ContentFilterErrorJSON    map[string]any `json:"content_filter_error_json,omitempty"`
 	ContentFilterErrorMessage string         `json:"content_filter_error_message,omitempty"`
 	ProbeMessage              string         `json:"probe_message,omitempty"`
-	RequestTimeout            *int           `json:"request_timeout,omitempty"`
+	RequestTimeout            *float64       `json:"request_timeout,omitempty"`
 }
 
 // TargetBackground holds target background context.
